moved members styles object to style_list object
